// models/wallet-type.ts
/**
 * Cobo Agentic Wallet Service
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */






/**
 * Supported wallet types.
 * @export
 * @enum {string}
 */
export const WalletType = {
    Custodial: 'Custodial' as const,
    MPC: 'MPC' as const,
} as const;

export type WalletType = typeof WalletType[keyof typeof WalletType];