// models/wallet-status.ts
/**
 * Cobo Agentic Wallet Service
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */






/**
 * Supported wallet statuses.
 * @export
 * @enum {string}
 */
export const WalletStatus = {
    preparing: 'preparing' as const,
    active: 'active' as const,
    archived: 'archived' as const,
} as const;

export type WalletStatus = typeof WalletStatus[keyof typeof WalletStatus];