// models/wallet-read.ts
/**
 * Cobo Agentic Wallet Service
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */





// @ts-ignore
import type { WalletStatus } from './wallet-status';

// @ts-ignore
import type { WalletType } from './wallet-type';




/**
 * Wallet response payload.
 * @export
 * @interface WalletRead
 */
export interface WalletRead {
    /**
     * The wallet ID.
     * @type {string}
     * @memberof WalletRead
     */
    'uuid': string;
    /**
     * The wallet type (for example, `MPC`).
     * @type {WalletType}
     * @memberof WalletRead
     */
    'wallet_type': WalletType;
    /**
     * The wallet name.
     * @type {string}
     * @memberof WalletRead
     */
    'name': string;
    /**
     * The ID of the principal that owns this wallet.
     * @type {string}
     * @memberof WalletRead
     */
    'owner_id': string;
    /**
     * 
     * @type {string}
     * @memberof WalletRead
     */
    'cobo_wallet_id': string | null;
    /**
     * The wallet status. Possible values: `preparing`, `active`, `archived`.
     * @type {WalletStatus}
     * @memberof WalletRead
     */
    'status': WalletStatus;
    /**
     * Arbitrary key-value metadata associated with the wallet.
     * @type {{ [key: string]: any; }}
     * @memberof WalletRead
     */
    'metadata': { [key: string]: any; };
    /**
     * The timestamp when the wallet was created, in ISO 8601 format.
     * @type {string}
     * @memberof WalletRead
     */
    'created_at': string;
    /**
     * The timestamp when the wallet was last updated, in ISO 8601 format.
     * @type {string}
     * @memberof WalletRead
     */
    'updated_at': string;
    /**
     * TSS node ID of the wallet owner\'s main device. Only set for MPC wallets.
     * @type {string}
     * @memberof WalletRead
     */
    'main_node_id'?: string;
}